functions/requirements.txt
azure-functions==1.18.0
azure-storage-queue==12.9.0
azure-messaging-webpubsubservice==1.2.2
psycopg[binary]==3.2.9
pydantic==2.7.3